Develop

+++ plugins/CouchDB.js
@@ -456,7 +456,17 @@ async function couchdb(fastify, options) {
               const res = [];
               if (format === 'summary') {
                 for (let i = 0; i < body.rows.length; i += 1) {
-                  // not putting project id. getprojectaims puts the project that was called from
+                  // If there is a project in the query, just send that back.
+                  // If there is no project in the search, body.rows[i].fields.project might return an array
+                  //   and the projectName retrieval from the map would fail.
+                  //   We are not handling that situation yet!
+                  let projectId = body.rows[i].fields.project;
+                  if (Array.isArray(projectId)) {
+                    const regex = /(?:project:")(\w+)(?:")/gm;
+                    const projectInQry = searchQry.q.match(regex);
+                    if (projectInQry && projectInQry[0])
+                      projectId = projectInQry[0].split(':')[1].replaceAll('"', '');
+                  }
                   res.push({
                     aimID: body.rows[i].id,
                     subjectID: body.rows[i].fields.patient_id,
@@ -479,8 +489,8 @@ async function couchdb(fastify, options) {
                     userName: Array.isArray(body.rows[i].fields.user)
                       ? body.rows[i].fields.user
                       : [body.rows[i].fields.user],
-                    projectID: body.rows[i].fields.project,
-                    projectName: projectNameMap[body.rows[i].fields.project],
+                    projectID: projectId,
+                    projectName: projectNameMap[projectId],
                     modality: body.rows[i].fields.modality,
                     anatomy: body.rows[i].fields.anatomy,
                     observation: body.rows[i].fields.observation,
@@ -558,7 +568,9 @@ async function couchdb(fastify, options) {
                             .then((result) => {
                               fastify.log.info(`Zip file ready in ${result}`);
                               // get the protocol and hostname from the request
-                              const link = `${request.protocol}://${request.hostname}${result}`;
+                              const link = `${config.httpsLink ? 'https' : request.protocol}://${
+                                request.hostname
+                              }${result}`;
                               // send notification and/or email with link
                               if (request)
                                 new EpadNotification(request, 'Download ready', link, false).notify(
@@ -1368,16 +1380,25 @@ async function couchdb(fastify, options) {
               if (!couchDoc.projects.includes(projectId)) couchDoc.projects.push(projectId);
             } else couchDoc.projects = [projectId];
           }
-
-          db.multipart
-            .insert(couchDoc, attachments, couchDoc._id)
-            .then(() => {
-              // await fastify.getAimVersions(couchDoc._id);
-              resolve(`Aim ${couchDoc._id} is saved successfully`);
-            })
-            .catch((err) => {
-              reject(new InternalError(`Saving aim ${couchDoc._id} to couchdb`, err));
-            });
+          if (attachments && attachments.length > 0)
+            db.multipart
+              .insert(couchDoc, attachments, couchDoc._id)
+              .then(() => {
+                // await fastify.getAimVersions(couchDoc._id);
+                resolve(`Aim ${couchDoc._id} is saved successfully`);
+              })
+              .catch((err) => {
+                reject(new InternalError(`Saving aim ${couchDoc._id} to couchdb`, err));
+              });
+          else
+            db.insert(couchDoc, couchDoc._id)
+              .then(() => {
+                // await fastify.getAimVersions(couchDoc._id);
+                resolve(`Aim ${couchDoc._id} is saved successfully`);
+              })
+              .catch((err) => {
+                reject(new InternalError(`Saving aim ${couchDoc._id} to couchdb`, err));
+              });
         });
       })
   );
@@ -1450,7 +1471,7 @@ async function couchdb(fastify, options) {
     (aimuid) =>
       new Promise((resolve, reject) => {
         const db = fastify.couch.db.use(config.db);
-        db.get(aimuid, (error, existing) => {
+        db.get(aimuid, async (error, existing) => {
           if (error || !existing) {
             reject(new ResourceNotFoundError('Aim', aimuid));
           }
@@ -1462,11 +1483,24 @@ async function couchdb(fastify, options) {
                 .segmentationEntityCollection;
             // this is a segmentation aim
             if (segEntity) {
-              const params = {
-                study: segEntity.SegmentationEntity[0].studyInstanceUid.root,
-                series: segEntity.SegmentationEntity[0].seriesInstanceUid.root,
-              };
-              promisses.push(fastify.deleteSeriesDicomsInternal(params));
+              // check if there are any other aims pointing to the DSO
+              // do we need to if we will always have only one aim pointing to the seg? what if in another project
+              // eslint-disable-next-line no-await-in-loop
+              const existingAim = await fastify.checkProjectSegAimExistence(
+                segEntity.SegmentationEntity[0].seriesInstanceUid.root,
+                undefined,
+                aimuid
+              );
+              if (!existingAim) {
+                const params = {
+                  study: segEntity.SegmentationEntity[0].studyInstanceUid.root,
+                  series: segEntity.SegmentationEntity[0].seriesInstanceUid.root,
+                };
+                promisses.push(fastify.deleteSeriesDicomsInternal(params));
+              } else
+                fastify.log.warn(
+                  `In aim system delete, Aim ${aimuid} refers to a segmentation with DSO Series UID ${segEntity.SegmentationEntity[0].seriesInstanceUid.root}. However, the DSO is referred by another aim ${existingAim}. It won't be deleted from the system`
+                );
             }
           }
